+++ src/components/dynamic/BooleanComponent.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from "react-i18next";
 import { FormGroup, Switch } from "@patternfly/react-core";
 
 import type { ComponentProps } from "./components";
-import { HelpItem } from "../../../components/help-enabler/HelpItem";
+import { HelpItem } from "../help-enabler/HelpItem";
 import { convertToHyphens } from "../../../util";
 
 export const BooleanComponent = ({
