+++ src/components/dynamic/ClientSelectComponent.tsx
@@ -8,20 +8,46 @@ import {
   SelectVariant,
 } from "@patternfly/react-core";
 
+import type { ClientQuery } from "@keycloak/keycloak-admin-client/lib/resources/clients";
+import { useAdminClient, useFetch } from "../../context/auth/AdminClient";
+import { HelpItem } from "../help-enabler/HelpItem";
 import type { ComponentProps } from "./components";
-import { HelpItem } from "../../../components/help-enabler/HelpItem";
 import { convertToHyphens } from "../../../util";
 
-export const ListComponent = ({
+export const ClientSelectComponent = ({
   name,
   label,
   helpText,
   defaultValue,
-  options,
 }: ComponentProps) => {
   const { t } = useTranslation("client-scopes");
   const { control } = useFormContext();
+
   const [open, setOpen] = useState(false);
+  const [clients, setClients] = useState<JSX.Element[]>();
+  const [search, setSearch] = useState("");
+
+  const adminClient = useAdminClient();
+
+  useFetch(
+    () => {
+      const params: ClientQuery = {
+        max: 20,
+      };
+      if (search) {
+        params.clientId = search;
+        params.search = true;
+      }
+      return adminClient.clients.find(params);
+    },
+    (clients) =>
+      setClients(
+        clients.map((option) => (
+          <SelectOption key={option.id} value={option.clientId} />
+        ))
+      ),
+    [search]
+  );
 
   return (
     <FormGroup
@@ -33,29 +59,26 @@ export const ListComponent = ({
     >
       <Controller
         name={`config.${convertToHyphens(name!)}`}
-        data-testid={name}
         defaultValue={defaultValue || ""}
         control={control}
         render={({ onChange, value }) => (
           <Select
             toggleId={name}
-            onToggle={(toggle) => setOpen(toggle)}
+            variant={SelectVariant.typeahead}
+            onToggle={(open) => setOpen(open)}
+            isOpen={open}
+            selections={value}
+            onFilter={(_, value) => {
+              setSearch(value);
+              return clients;
+            }}
             onSelect={(_, value) => {
-              onChange(value as string);
+              onChange(value.toString());
               setOpen(false);
             }}
-            selections={value}
-            variant={SelectVariant.single}
             aria-label={t(label!)}
-            isOpen={open}
           >
-            {options!.map((option) => (
-              <SelectOption
-                selected={option === value}
-                key={option}
-                value={option}
-              />
-            ))}
+            {clients}
           </Select>
         )}
       />
