+++ src/components/dynamic/ListComponent.tsx
@@ -8,46 +8,20 @@ import {
   SelectVariant,
 } from "@patternfly/react-core";
 
-import type { ClientQuery } from "@keycloak/keycloak-admin-client/lib/resources/clients";
-import { useAdminClient, useFetch } from "../../../context/auth/AdminClient";
-import { HelpItem } from "../../../components/help-enabler/HelpItem";
 import type { ComponentProps } from "./components";
+import { HelpItem } from "../help-enabler/HelpItem";
 import { convertToHyphens } from "../../../util";
 
-export const ClientSelectComponent = ({
+export const ListComponent = ({
   name,
   label,
   helpText,
   defaultValue,
+  options,
 }: ComponentProps) => {
   const { t } = useTranslation("client-scopes");
   const { control } = useFormContext();
-
   const [open, setOpen] = useState(false);
-  const [clients, setClients] = useState<JSX.Element[]>();
-  const [search, setSearch] = useState("");
-
-  const adminClient = useAdminClient();
-
-  useFetch(
-    () => {
-      const params: ClientQuery = {
-        max: 20,
-      };
-      if (search) {
-        params.clientId = search;
-        params.search = true;
-      }
-      return adminClient.clients.find(params);
-    },
-    (clients) =>
-      setClients(
-        clients.map((option) => (
-          <SelectOption key={option.id} value={option.clientId} />
-        ))
-      ),
-    [search]
-  );
 
   return (
     <FormGroup
@@ -59,26 +33,29 @@ export const ClientSelectComponent = ({
     >
       <Controller
         name={`config.${convertToHyphens(name!)}`}
+        data-testid={name}
         defaultValue={defaultValue || ""}
         control={control}
         render={({ onChange, value }) => (
           <Select
             toggleId={name}
-            variant={SelectVariant.typeahead}
-            onToggle={(open) => setOpen(open)}
-            isOpen={open}
-            selections={value}
-            onFilter={(_, value) => {
-              setSearch(value);
-              return clients;
-            }}
+            onToggle={(toggle) => setOpen(toggle)}
             onSelect={(_, value) => {
-              onChange(value.toString());
+              onChange(value as string);
               setOpen(false);
             }}
+            selections={value}
+            variant={SelectVariant.single}
             aria-label={t(label!)}
+            isOpen={open}
           >
-            {clients}
+            {options!.map((option) => (
+              <SelectOption
+                selected={option === value}
+                key={option}
+                value={option}
+              />
+            ))}
           </Select>
         )}
       />
